Add and import helper functions
