+++ tests/test_api.py
@@ -1,4 +1,4 @@
-from ride_my_way import views
+from ride_my_way import helpers, views
 import unittest
 import json
 import jwt
@@ -26,8 +26,8 @@ class TestRideMyWay(unittest.TestCase):
     def setUp(self):
         # views.app.config['TESTING'] = True
         self.app = views.app.test_client()
-        self.cur = views.database_connection
-        views.database_connection.create_tables()
+        self.cur = helpers.database_connection
+        helpers.database_connection.create_tables()
 
         # --------***** Creating users ********------------------
